nvim: bug fix, pyexec latex
for some reason the minted_environment's (code_block) parent's next
sibling is the paragraph node (Output), instead of being the next
sibling of minted_environment, only in some scenerios. Added a check in
_add_output_block_to_tex() and _remove_output_block_from_tex() for catching
this case.

if we won't catch this the, pyexec will add another (or won't remove)
Output block even if there's one right next to the code_block
(minted_environment).